Tidy some types and names.

+++ cpp/dolfin/mesh/Connectivity.h
@@ -81,10 +81,10 @@ public:
   Connectivity& operator=(Connectivity&& connectivity) = default;
 
   /// Return number of connections for given entity
-  std::size_t size(std::int32_t entity) const;
+  std::int32_t size(std::int32_t entity) const;
 
   /// Return global number of connections for given entity
-  std::size_t size_global(std::int32_t entity) const;
+  std::int64_t size_global(std::int32_t entity) const;
 
   /// Return array of connections for given entity
   std::int32_t* connections(int entity);
